Move renameInCreateQuery() to a better place.
