+++ src/Backups/RestoreFromBackupUtils.cpp
@@ -1,10 +1,9 @@
 #include <Backups/BackupUtils.h>
-#include <Backups/BackupRenamingConfig.h>
+#include <Backups/DDLRenamingVisitor.h>
 #include <Backups/IBackup.h>
 #include <Backups/IBackupEntry.h>
 #include <Backups/IRestoreFromBackupTask.h>
 #include <Backups/hasCompatibleDataToRestoreTable.h>
-#include <Backups/renameInCreateQuery.h>
 #include <Common/escapeForFileName.h>
 #include <Databases/IDatabase.h>
 #include <IO/ReadHelpers.h>
@@ -175,10 +174,8 @@ namespace
         /// Prepares internal structures for making tasks for restoring.
         void prepare(const ASTBackupQuery::Elements & elements)
         {
-            auto new_renaming_config = std::make_shared<BackupRenamingConfig>();
             String current_database = context->getCurrentDatabase();
-            new_renaming_config->setFromBackupQueryElements(elements, current_database);
-            renaming_config = new_renaming_config;
+            renaming_settings.setFromBackupQuery(elements, current_database);
 
             for (const auto & element : elements)
             {
@@ -238,7 +235,7 @@ namespace
         void prepareToRestoreTable(const DatabaseAndTableName & table_name_, const ASTs & partitions_)
         {
             /// Check that we are not trying to restore the same table again.
-            DatabaseAndTableName new_table_name = renaming_config->getNewTableName(table_name_);
+            DatabaseAndTableName new_table_name = renaming_settings.getNewTableName(table_name_);
             if (tables.contains(new_table_name))
             {
                 String message;
@@ -310,7 +307,7 @@ namespace
         void prepareToRestoreDatabase(const String & database_name_, const std::set<String> & except_list_)
         {
             /// Check that we are not trying to restore the same database again.
-            String new_database_name = renaming_config->getNewDatabaseName(database_name_);
+            String new_database_name = renaming_settings.getNewDatabaseName(database_name_);
             if (databases.contains(new_database_name) && databases[new_database_name].is_explicit)
                 throw Exception(ErrorCodes::CANNOT_RESTORE_DATABASE, "Couldn't restore database {} twice", backQuoteIfNeed(new_database_name));
 
@@ -407,7 +404,7 @@ namespace
         /// Do renaming in the create query according to the renaming config.
         std::shared_ptr<ASTCreateQuery> renameInCreateQuery(const ASTPtr & ast) const
         {
-            return typeid_cast<std::shared_ptr<ASTCreateQuery>>(::DB::renameInCreateQuery(ast, renaming_config, context));
+            return typeid_cast<std::shared_ptr<ASTCreateQuery>>(::DB::renameInCreateQuery(ast, context, renaming_settings));
         }
 
         static bool isSystemOrTemporaryDatabase(const String & database_name)
@@ -442,7 +439,7 @@ namespace
 
         ContextMutablePtr context;
         BackupPtr backup;
-        BackupRenamingConfigPtr renaming_config;
+        DDLRenamingSettings renaming_settings;
         std::map<String, CreateDatabaseInfo> databases;
         std::map<DatabaseAndTableName, CreateTableInfo> tables;
     };
