+++ src/Common/ErrorCodes.cpp
@@ -614,6 +614,7 @@
     M(643, CANNOT_UNPACK_ARCHIVE) \
     M(644, CANNOT_BACKUP_DATABASE) \
     M(645, CANNOT_BACKUP_TABLE) \
+    M(646, WRONG_DDL_RENAMING_SETTINGS) \
     \
     M(999, KEEPER_EXCEPTION) \
     M(1000, POCO_EXCEPTION) \
